Pass witness_type directly to uint64 depth API.
Remove the obsolete int32_t intermediate cast in stress_test so the call matches the updated verifier depth parameter type.

+++ src/stress_test.cpp
@@ -40,7 +40,7 @@ void doit(int thread, std::vector<job> const& jobs)
             j.outputbytes.data(),
             j.outputbytes.size(),
             j.number_of_iterations,
-            static_cast<int32_t>(j.witness_type));
+            j.witness_type);
         if (!is_valid) {
             printf("thread %d cnt %d is valid %d %" PRIu64 " %d\n",
                 thread,
